service/requirements.txt
psycopg2-binary==2.9.9
uvicorn==0.30.3
fastapi==0.111.1
sqlalchemy==2.0.31
python-dotenv==1.0.1
requests==2.32.3
